Fix: gradients, textures and animations now apply in both modes
The previous version put the background, texture and animation on the
Board's outer Stack, and rows fell back to bg.from when no per-line
colour was set — so the rows painted opaque over the gradient/texture
and hid every effect except the row colour itself. Photo mode bypassed
all of it because BigPicture just rendered an image with no overlay.

Hoist the visual layers into a new Stage component that wraps both the
Board and BigPicture:

- Stage owns the gradient background, texture overlay, texture animation,
  and the shimmer pass.
- Board only renders rows. Rows default to transparent; an explicit
  Line.color is the only thing that paints over the Stage.
- BigPicture renders just the image (or nothing if none uploaded), so
  the Stage's gradient and texture show through around contained images
  and behind a missing one.

Result: gradients, dots/stripes/grid/noise, pan/pulse/shimmer all
render the same in normal mode and photo mode.
